Add ApiGateway service and update configurations for services
Added ApiGateway microservice to the project with bases like JWT authentication, Ocelot configuration, and Swagger for API documentation. Updated docker and launch configurations for the ApiGateway to work with other updated services. Also updated catalogue and basket services to .NET 8.0. Introduced a GRPC connection in the basket service to communicate with discount service.