Add new projects and references in Ordering service
Three new projects (Ordering.Application, Ordering.Domain, Ordering.Infrastructure) have been added to the Ordering service. References to these projects are now included in the Ordering.API project as well as in the solution file ('CursoMicroservices.sln'). This commit organizes the architecture of the Ordering service for modularity and separation of concerns.